ml/ranker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.3
google-cloud-bigquery
numpy
xgboost
